+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### 0.0.1
 
-- Proof of concept library
+- proof of concept library
 - implementation of `resurrect` decorator
 
 ### 0.0.2
@@ -13,3 +13,7 @@
 - improved logging
 - new short term memory functions: save memory to a json file, get current memory content, reset current memory content
 - implementation of `mutate` decorator
+
+### 0.0.3
+
+- added `allow_installs` argument to `mutate` and `resurrect` decorator. Allows LLM to install non installed python libraries